Literal {tables}	R Documentation

Insert a literal entry into a table margin.

Description

This allows insertion of arbitrary LaTeX text into a table.

Usage

Literal(x, nearData = TRUE)

Arguments

x
A character string to insert.

nearData
See the Details section of Heading.

Details

In LaTeX the literal string should usually end with a % comment character to avoid having a blank line inserted.

Value

Produces an expression to insert a label containing the literal text.

See Also

Hline, which uses this to insert lines.

Examples

tabular( (Literal("Some text") + Species)  ~ 
       All(iris)*mean, data=iris )

[Package tables version 0.9.17 Index]