Pseudo-function to compute a statistic relative to a reference set.

The Percent pseudo-function is used to specify a statistic that depends on other values in the table.

Usage

Percent(denom = "all", fn = percent)

Arguments

denom

How the reference set (the denominator in case of a percentage calculation) should be calculated. See below.

fn

The two argument function to calculate the statistic.

Details

The function fn will be called with two arguments. The first argument is the usual “value vector” of values corresponding to this cell in the table, and the second is another vector of reference values, determined by denom.

The default value of fn is the percent function, defined as function(x, y) 100*length(x)/length(y). This gives the ratio of the number of values in the current cell relative to the reference values, expressed as a percentage. Using fn = function(x, y) 100*sum(x)/sum(y) would give the percentage of the sum of the values in the current cell to the sum in the reference set.

With the default denom = "all", all values of the analysis variable in the dataset are used as the reference. Other possibilities are denom = "row" or denom = "col", for which the values of the variable corresponding to the current row or column subset are used.

The special syntax denom = Equal(...) will record each expression in .... The reference set will be the cases with equal values of all expressions in ... to those of the current cell. The similar form denom = Unequal(...) sets the reference values to be those that differ in any of the ... expressions from the current cell. (In fact, these can be used somewhat more generally; see the vignette for details.)

Finally, other possible denom values are a logical vector, in which case the values marked TRUE are used, or anything else, which will be passed to fn as y without any subsetting. (To pass a variable with subsetting, use the Arguments pseudo-function instead.)

Pseudo-functions

Percent is a “pseudo-function”: it takes the form of a function call, but is never actually called: it is handled specially by tabular. Equal and Unequal are also pseudo-functions, but are only special when used in the denom argument to Percent.

See also

Examples

x <- factor(sample(LETTERS[1:2], 1000, rep=TRUE))
y <- factor(sample(letters[3:4], 1000, rep=TRUE))
z <- factor(sample(LETTERS[5:6], 1000, rep=TRUE))

# These both do the same thing:
tabular( (x + 1)*(y + 1) ~ (z + 1)*(1+(RowPct=Percent("row"))))
#>                                           
#>          z                                
#>          E          F          All        
#>  x   y   All RowPct All RowPct All  RowPct
#>  A   c   111 44.76  137 55.24   248 100   
#>      d   127 52.05  117 47.95   244 100   
#>      All 238 48.37  254 51.63   492 100   
#>  B   c   115 45.28  139 54.72   254 100   
#>      d   123 48.43  131 51.57   254 100   
#>      All 238 46.85  270 53.15   508 100   
#>  All c   226 45.02  276 54.98   502 100   
#>      d   250 50.20  248 49.80   498 100   
#>      All 476 47.60  524 52.40  1000 100   
tabular( (x + 1)*(y + 1) ~ (z + 1)*(1+(xyPct=Percent(Equal(x, y)))))
#>                                        
#>          z                             
#>          E         F         All       
#>  x   y   All xyPct All xyPct All  xyPct
#>  A   c   111 44.76 137 55.24  248 100  
#>      d   127 52.05 117 47.95  244 100  
#>      All 238 48.37 254 51.63  492 100  
#>  B   c   115 45.28 139 54.72  254 100  
#>      d   123 48.43 131 51.57  254 100  
#>      All 238 46.85 270 53.15  508 100  
#>  All c   226 45.02 276 54.98  502 100  
#>      d   250 50.20 248 49.80  498 100  
#>      All 476 47.60 524 52.40 1000 100