as.tabular {tables}	R Documentation

Convert matrix or dataframe to tabular object.

Description

These functions construct or copy labels onto an existing matrix or dataframe.

Usage

as.tabular(x, like = NULL)
## Default S3 method:
as.tabular(x, like = NULL)
## S3 method for class 'data.frame'
as.tabular(x, like = NULL)

Arguments

x
The object to convert.

like
If not NULL, should be a tabular object with the same number of rows and columns as x. Its labels will be used on the result.

Value

A tabular object.

See Also

as.matrix.tabular

Examples

model <- tabular( (Species + 1) ~ (n=1) + Sepal.Length + Sepal.Width, data=iris )
model
as.tabular(matrix(1:12, 4,3), like=model)

[Package tables version 0.9.17 Index]