labelSubset {tables}	R Documentation

Add a label to a logical vector.

Description

This function is mainly for internal use. It adds a label to a logical vector, so that the Percent pseudo-function can ignore it when forming a denominator.

Usage

labelSubset(subset, label)

Arguments

subset
A logical vector describing a subset of the dataset.

label
A character label to use to describe this subset in a call to Equal or Unequal within Percent.

Value

A vector of class "labelledSubset" with the label recorded as an attribute.

Author(s)

Duncan Murdoch

See Also

Percent

[Package tables version 0.9.17 Index]