Hline {tables}	R Documentation

Add a horizontal line to a LaTeX table.

Description

This function inserts a LaTeX directive to draw a full or partial line in a table.

Usage

Hline(columns, nearData = FALSE)

Arguments

columns
Which columns should receive the line?

nearData
See the Details section of Heading.

Details

Hline() is not very flexible: it must be the leftmost header in a row specification for the table, i.e. mean * Hline() is not allowed. Anything to the right of the Hline() factor will be ignored.

Value

Produces an expression to insert a label which will be interpreted by LaTeX as a request for a horizontal line.

Examples

toLatex( tabular( Species + Hline() + 1 ~ mean*Sepal.Width, data=iris) )

[Package tables version 0.9.17 Index]