Justify {tables}	R Documentation

Justify pseudo-function

Description

The Justify pseudo-function sets the justification of the following items in the table.

Usage

Justify(labels, data=labels)

Arguments

labels
Justification to use for labels

data
Justification to use for data.

Details

The justification can be an R name if that is syntactically valid, or a quoted string.

Pseudo-functions

This is a “pseudo-function”: it takes the form of a function call, but is never actually called: it is handled specially by tabular.

Examples

tabular( Justify(c,l)*Heading(Var)*(Sepal.Length+Sepal.Width) ~ 
         Justify(c)*(mean + sd), data=iris )

[Package tables version 0.9.17 Index]