labels {tables}	R Documentation

Retrieve or modify the row or column labels.

Description

These functions allow the row or column labels of a tabular object to be retrieved or modified.

Usage

rowLabels(x)
rowLabels(x) <- value
colLabels(x)
colLabels(x) <- value
## S3 method for class 'tabularRowLabels'
x[i, j, ..., drop = FALSE]
## S3 method for class 'tabularColLabels'
x[i, j, ..., drop = FALSE]

Arguments

x
A "tabular", "tabularRowLabels" or "tabularColLabels" object.

value
A replacement

i, j, ..., drop
Arguments used for subsetting the labels. See Details below.

Details

Subsetting the row labels does not allow the number of rows to be changed; likewise, subsetting the column labels does not allow the number of columns to be changed. To change both, subset the original "tabular" object.

Value

rowLabels and the corresponding subsetting method return an object of class "tabularRowLabels".

colLabels and the corresponding subsetting method return an object of class "tabularColLabels".

The assignment functions return "tabular" objects.

See Also

[.tabular

Examples

tab <- tabular( (Species + 1) ~ (n=1) + Format(digits=2)*
         (Sepal.Length + Sepal.Width)*(mean + sd), data=iris )
colLabels(tab)
colLabels(tab) <- colLabels(tab)[1,]
tab

[Package tables version 0.9.17 Index]